Clean up of codebase to remove unused functions: all stored in Exhaustive_search_depreceated.py
