+++ select_occupancy_groups.py
@@ -1,13 +1,14 @@
 from __future__ import print_function
-from iotbx.pdb import hierarchy
-import iotbx
-import libtbx.phil
+
 import sys
-from giant.structure.restraints.occupancy import overlapping_occupancy_groups
-from giant.maths.geometry import pairwise_dists
-from select_hierarchies import hierarchy_all_overlap
-import itertools
 from copy import deepcopy
+
+import iotbx
+import itertools
+import libtbx.phil
+from giant.maths.geometry import pairwise_dists
+from giant.structure.restraints.occupancy import overlapping_occupancy_groups
+from iotbx.pdb import hierarchy
 
 blank_arg_prepend = {'.pdb': 'pdb=', '.mtz': 'mtz=', '.csv': 'csv='}
 ##############################################################
@@ -243,24 +244,6 @@ def process_refined_pdb_bound_ground_states(pdb, params=master_phil.extract()):
 
     return bound_states, ground_states
 
-# def get_largest_concident_group(pdb, residue_altloc_dict, params):
-#
-#     coincident = get_coincident_group(pdb, residue_altloc_dict, params)
-#
-#     #print(residue_altloc_dict)
-#
-#     for residue_chain in residue_altloc_dict.keys():
-#         residue = residue_chain[0]
-#         chain = residue_chain[1]
-#         altloc_group = residue_altloc_dict.get(residue_chain)
-#         non_coincident_altlocs = list(altloc_group)
-#         for conincident_altloc_group in coincident:
-#             if conincident_altloc_group[1] == residue and conincident_altloc_group[2] == chain:
-#                 for altloc in conincident_altloc_group[0]:
-#                     non_coincident_altlocs.remove(altloc)
-#         if non_coincident_altlocs:
-#             raise ValueError("More than two groups of altlocs are present. This is currently not handled")
-
 def get_ligand_coincident_altloc_group(hier, coincident, params):
 
     resnames = params.occupancy.resnames.split(',')
@@ -283,143 +266,6 @@ def get_ligand_coincident_altloc_group(hier, coincident, params):
              return coincident_altloc_group[0]
         else:
             continue
-
-######################################################################
-# Selection using altlocs where all residues are shared, and they overlap
-######################################################################
-
-# Use occupancy groups to generate selection of atoms associated to bound & ground structure
-
-# Return an atom selection on the protein hier to be used as selection
-# Provide whether sufficent for co-varying
-
-# Need to consider generation of grid (for summing FoFc) for seperated components.
-
-###########################################################################
-# Used in generation of cartesian point selection
-###########################################################################
-def generate_altloc_hiearchy(altloc_residue_dict, pdb):
-
-    pdb_in = hierarchy.input(pdb)
-    altloc_selection = generate_altloc_selection(altloc_residue_dict, pdb)
-    altloc_hierarchy = pdb_in.hierarchy.select(altloc_selection)
-
-    print_hier_atoms(altloc_hierarchy)
-
-    return altloc_hierarchy
-
-def generate_altloc_selection(altloc_residue_dict, pdb):
-
-    pdb_in = hierarchy.input(pdb)
-    sel_cache = pdb_in.hierarchy.atom_selection_cache()
-
-    selection_list = []
-    for set in altloc_residue_dict.values():
-        for resseq_id in set:
-            selection_list.append("resid {} or ".format(resseq_id))
-
-    selection_string = ''.join(selection_list)
-    selection_string = selection_string.rstrip(" or ")
-    selection_string = "altloc {} and ({})".format(str(altloc_residue_dict.keys()[0]), selection_string)
-
-    altloc_selection = sel_cache.selection(selection_string)
-
-    return altloc_selection
-
-def from_altloc_generate_altloc_selection(pdb, altloc, occupancy_groups, params=master_phil.extract()):
-
-    altloc_residue_dict = get_altloc_residue_dict(altloc, occupancy_groups)
-    altloc_selection = generate_altloc_selection(altloc_residue_dict, pdb)
-
-    return altloc_selection
-
-def get_coincident_altloc_groups(altloc_groups_with_shared_residues, occupancy_groups, pdb):
-
-    coincident_altloc_groups = []
-    for altloc_group in altloc_groups_with_shared_residues:
-        print("Altlocs with Shared residues {}".format(altloc_group))
-
-        all_altloc_hier = []
-        for altloc in altloc_group:
-            altloc_residue_dict = get_altloc_residue_dict(altloc, occupancy_groups)
-            altloc_hierarchy = generate_altloc_hiearchy(altloc_residue_dict, pdb)
-            all_altloc_hier.append(altloc_hierarchy)
-
-        # TODO Rewrite hierarchy overlap to handle any number of hierarchies
-
-        if len(all_altloc_hier) > 2:
-            raise Warning("Only two hierarchies can be compared currently. "
-                          "Assuming that if a larger group, that they all overlap if the first two states overlap")
-
-        if hierarchy_all_overlap(all_altloc_hier[0], all_altloc_hier[1], rmsd_distance=0.02):
-            print("Altlocs {} are Coincident hierarchies".format(altloc_group))
-            coincident_altloc_groups.append(altloc_group)
-
-    return coincident_altloc_groups
-
-def get_altloc_residue_dict(altlocs,occupancy_groups):
-
-    altloc_residue_dict = dict()
-    for altloc in altlocs:
-        altloc_residue_set = set()
-        for occupancy_group in occupancy_groups:
-            for group in occupancy_group:
-                for residue_altloc in group:
-                    if altloc == residue_altloc.get("altloc"):
-                        altloc_residue_set.add(residue_altloc.get("resseq"))
-                        altloc_residue_dict.update({altloc: altloc_residue_set})
-
-    return altloc_residue_dict
-
-
-def find_altlocs_with_shared_residues(altlocs, occupancy_groups):
-
-    altloc_residue_dict = get_altloc_residue_dict(altlocs,occupancy_groups)
-
-    altloc_groups_with_shared_residues = list()
-    for altloc_group in powerset(altloc_residue_dict.keys(), min_len=2):
-        altloc_group_set = set()
-        for altloc in altloc_group:
-            altloc_group_set = altloc_group_set.union(altloc_residue_dict.get(altloc))
-
-        different_residues_in_set = False
-        for altloc in altloc_group:
-            if altloc_group_set.difference(altloc_residue_dict.get(altloc)):
-                different_residues_in_set = True
-            else:
-                pass
-        if not different_residues_in_set:
-            altloc_groups_with_shared_residues.append(altloc_group)
-
-    return (altloc_groups_with_shared_residues)
-
-# TODO Clean up how this sources the params
-def get_coincident_altlocs(pdb, params=master_phil.extract()):
-
-    occupancy_groups = get_occupancy_groups(pdb, params)
-    altlocs = get_parameter_from_occupancy_groups(occupancy_groups, "altloc")
-    altloc_groups_with_shared_residues = find_altlocs_with_shared_residues(altlocs, occupancy_groups)
-
-    if len(altloc_groups_with_shared_residues) > 2:
-        raise Warning("Current code can only handle two states")
-
-    coincident_altloc_groups = get_coincident_altloc_groups(altloc_groups_with_shared_residues,
-                                                            occupancy_groups, pdb)
-
-    print ("Coincident_altloc_groups: {}".format(coincident_altloc_groups))
-
-    return coincident_altloc_groups
-
-# TODO Clean up how this sources the params
-def get_altloc_hier(pdb, params=master_phil.extract()):
-    coincident_altloc_groups = get_coincident_altlocs(pdb, params)
-    occupancy_groups = get_occupancy_groups(pdb, params)
-
-    for altloc_group in coincident_altloc_groups:
-        altloc_residue_dict = get_altloc_residue_dict(altloc_group[0], occupancy_groups)
-        altloc_hier = generate_altloc_hiearchy(altloc_residue_dict, pdb)
-
-        yield altloc_hier
 
 #######################################################################
 # Helper functions
